mfulc_reader_sim/pyproject.toml
[project]
name = "mfulc-reader-sim"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pycryptodome>=3.23.0",
]
